solved mypy errors

+++ src/ghg_forcing_for_cmip_comparison/utils.py
@@ -100,20 +100,20 @@ class GroundDataSchema(pa.DataFrameModel):
     time: Series[pd.DatetimeTZDtype] = pa.Field(
         dtype_kwargs={"unit": "ns", "tz": "UTC"}
     )
-    year: Series[np.int64] = pa.Field(ge=1968, le=datetime.now().year)
-    month: Series[np.int64] = pa.Field(ge=1, le=12)
-    latitude: Series[np.float64] = pa.Field(ge=-90, le=90)
-    longitude: Series[np.float64] = pa.Field(ge=-180, le=180)
-    lat_bnd: Series[np.int64] = pa.Field(ge=-90, le=90)
-    lon_bnd: Series[np.int64] = pa.Field(ge=-180, le=180)
-    lat: Series[np.float64] = pa.Field(gt=-90, lt=90)
-    lon: Series[np.float64] = pa.Field(gt=-180, lt=180)
-    value: Series[np.float64] = pa.Field(gt=0.0, lt=9999.0)
-    std_dev: Series[np.float64] = pa.Field(ge=0)
-    numb: Series[np.int64] = pa.Field(gt=0)
+    year: Series[int] = pa.Field(ge=1968, le=datetime.now().year)
+    month: Series[int] = pa.Field(ge=1, le=12)
+    latitude: Series[float] = pa.Field(ge=-90, le=90)
+    longitude: Series[float] = pa.Field(ge=-180, le=180)
+    lat_bnd: Series[int] = pa.Field(ge=-90, le=90)
+    lon_bnd: Series[int] = pa.Field(ge=-180, le=180)
+    lat: Series[float] = pa.Field(gt=-90, lt=90)
+    lon: Series[float] = pa.Field(gt=-180, lt=180)
+    value: Series[float] = pa.Field(gt=0.0, lt=9999.0)
+    std_dev: Series[float] = pa.Field(ge=0)
+    numb: Series[int] = pa.Field(gt=0)
     site_code: Series[str]
     network: Series[str] = pa.Field(isin=["agage", "gage", "noaa"])
-    altitude: Series[np.float64] = pa.Field(ge=0.0, nullable=True)
+    altitude: Series[float] = pa.Field(ge=0.0, nullable=True)
     insitu_vs_flask: Series[str] = pa.Field(isin=["flask", "insitu"], nullable=True)
     sampling_strategy: Series[str] = pa.Field(
         isin=["surface-flask", "shipboard-flask", "surface-insitu"], nullable=True
@@ -131,21 +131,21 @@ class EODataSchema(pa.DataFrameModel):
     time: Series[pd.DatetimeTZDtype] = pa.Field(
         dtype_kwargs={"unit": "ns", "tz": "UTC"}
     )
-    year: Series[np.int64] = pa.Field(ge=1968, le=datetime.now().year)
-    month: Series[np.int64] = pa.Field(ge=1, le=12)
-    lat_bnd: Series[np.int64] = pa.Field(ge=-90, le=90)
-    lon_bnd: Series[np.int64] = pa.Field(ge=-180, le=180)
-    bnd: Series[np.int64] = pa.Field(isin=[0, 1])
-    lat: Series[np.float64] = pa.Field(gt=-90, lt=90)
-    lon: Series[np.float64] = pa.Field(gt=-180, lt=180)
-    value: Series[np.float64] = pa.Field(gt=0.0, lt=9999.0)
-    std_dev: Series[np.float64] = pa.Field(ge=0)
-    numb: Series[np.int64] = pa.Field(gt=0)
+    year: Series[int] = pa.Field(ge=1968, le=datetime.now().year)
+    month: Series[int] = pa.Field(ge=1, le=12)
+    lat_bnd: Series[int] = pa.Field(ge=-90, le=90)
+    lon_bnd: Series[int] = pa.Field(ge=-180, le=180)
+    bnd: Series[int] = pa.Field(isin=[0, 1])
+    lat: Series[float] = pa.Field(gt=-90, lt=90)
+    lon: Series[float] = pa.Field(gt=-180, lt=180)
+    value: Series[float] = pa.Field(gt=0.0, lt=9999.0)
+    std_dev: Series[float] = pa.Field(ge=0)
+    numb: Series[int] = pa.Field(gt=0)
     gas: Series[str] = pa.Field(isin=["ch4", "co2"])
     unit: Series[str] = pa.Field(isin=["ppb", "ppm"])
-    pre: Series[np.float64] = pa.Field(gt=0, lt=1)
-    column_averaging_kernel: Series[np.float64]
-    vmr_profile_apriori: Series[np.float64]
+    pre: Series[float] = pa.Field(gt=0, lt=1)
+    column_averaging_kernel: Series[float]
+    vmr_profile_apriori: Series[float]
 
 
 @task(name="save_dataset")
